Move file picking to its own utils submodule and add folder picking
