+++ ronald-egui/src/utils.rs
@@ -1,2 +1,3 @@
+pub mod files;
 pub mod sync;
 pub mod vectorize;
